+++ NewsArticle+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension NewsArticle {
         return NSFetchRequest<NewsArticle>(entityName: "NewsArticle")
     }
 
-    @NSManaged public var article_id: Int16
+    @NSManaged public var article_id: Int64
     @NSManaged public var blurb: String?
     @NSManaged public var category: String?
     @NSManaged public var category_id: Int16
